Split "user" into "account" and "role"
