+++ hprsctool/hprsctool.py
@@ -6,7 +6,8 @@ import sys
 from .commands.manager import manager
 from .commands import system
 from .commands import task
-from .commands import user
+from .commands import account
+from .commands import role
 from .comm import remote_system_controller
 
 VERSION = "0.11.0"
@@ -55,9 +56,13 @@ def main():
     task_subparsers = subparser.add_subparsers()
     task.get_parameters(task_subparsers)
 
-    subparser = top_level_subparsers.add_parser("user", help="User management commands")
-    user_subparsers = subparser.add_subparsers()
-    user.get_parameters(user_subparsers)
+    subparser = top_level_subparsers.add_parser("account", help="Account management commands")
+    account_subparsers = subparser.add_subparsers()
+    account.get_parameters(account_subparsers)
+
+    subparser = top_level_subparsers.add_parser("role", help="Role management commands")
+    role_subparsers = subparser.add_subparsers()
+    role.get_parameters(role_subparsers)
 
     args = argparser.parse_args(remaining_args)
     config = remote_system_controller.RedfishConfig(
